mscore::algorithm::isotope

Function generate_averagine_spectrum

pub fn generate_averagine_spectrum(
    mass: f64,
    charge: i32,
    min_intensity: i32,
    k: i32,
    resolution: i32,
    centroid: bool,
    amp: Option<f64>,
) -> MzSpectrum

generate the averagine spectrum for a given mass and charge

Arguments:

mass - mass of the peptide

charge - charge of the peptide

min_intensity - minimum intensity for a peak to be included in the result

k - number of isotopes to consider

resolution - resolution of the isotope pattern

centroid - whether to centroid the spectrum

amp - amplitude of the isotope pattern

Returns:

MzSpectrum - averagine spectrum

Examples

use mscore::algorithm::isotope::generate_averagine_spectrum;

let spectrum = generate_averagine_spectrum(3000.0, 2, 1, 10, 3, true, None);