mscore::algorithm

Module peptide

Functions

atomic_product_ion_composition

calculate the atomic composition of a product ion

calculate_amino_acid_composition

get a count dictionary of the amino acid composition of a peptide sequence

calculate_complementary_fragment_composition

Calculate the atomic composition of the complementary fragment.

calculate_peptide_mono_isotopic_mass

calculate the monoisotopic mass of a peptide sequence

calculate_peptide_product_ion_mono_isotopic_mass

calculate the monoisotopic mass of a peptide product ion for a given fragment type

calculate_product_ion_mz

calculate the monoisotopic m/z of a peptide product ion for a given fragment type and charge

fragments_to_composition

calculate the atomic composition of a peptide product ion series Arguments:

get_num_protonizable_sites

count the number of protonizable sites in a peptide sequence

peptide_sequence_to_atomic_composition

calculate the atomic composition of a peptide sequence

simulate_charge_state_for_sequence

simulate the charge state distribution for a peptide sequence

simulate_charge_states_for_sequences

simulate the charge state distribution for a list of peptide sequences