mscore::algorithm::peptide

Function get_num_protonizable_sites

pub fn get_num_protonizable_sites(sequence: &str) -> usize

count the number of protonizable sites in a peptide sequence

Arguments

sequence - a string representing the peptide sequence

Returns

usize - the number of protonizable sites in the peptide sequence

Example

use mscore::algorithm::peptide::get_num_protonizable_sites;

let sequence = "PEPTIDEH";
let num_protonizable_sites = get_num_protonizable_sites(sequence);
assert_eq!(num_protonizable_sites, 2);