mscore::chemistry

Module utility

Functions

find_unimod_patterns

Convert a peptide sequence with UNIMOD annotations to a tuple of plain sequence and for each position in the sequence, the mass of the modification at that position (0 if no modification), which is the representation of sequence nad modifications used by SAGE

reshape_prosit_array

Reshape the flat prosit array into a 3D array of shape (29, 2, 3)

unimod_sequence_to_tokens

Convert a peptide sequence with UNIMOD annotations to a list of tokens