Trait mscore::timstof::quadrupole::IonTransmission

pub trait IonTransmission {
    // Required method
    fn apply_transmission(
        &self,
        frame_id: i32,
        scan_id: i32,
        mz: &Vec<f64>,
    ) -> Vec<f64>;

    // Provided methods
    fn transmit_spectrum(
        &self,
        frame_id: i32,
        scan_id: i32,
        spectrum: MzSpectrum,
        min_probability: Option<f64>,
    ) -> MzSpectrum { ... }
    fn transmit_annotated_spectrum(
        &self,
        frame_id: i32,
        scan_id: i32,
        spectrum: MzSpectrumAnnotated,
        min_probability: Option<f64>,
    ) -> MzSpectrumAnnotated { ... }
    fn transmit_ion(
        &self,
        frame_ids: Vec<i32>,
        scan_ids: Vec<i32>,
        spec: MzSpectrum,
        min_proba: Option<f64>,
    ) -> Vec<Vec<MzSpectrum>> { ... }
    fn get_transmission_set(
        &self,
        frame_id: i32,
        scan_id: i32,
        mz: &Vec<f64>,
        min_proba: Option<f64>,
    ) -> HashSet<usize> { ... }
    fn all_transmitted(
        &self,
        frame_id: i32,
        scan_id: i32,
        mz: &Vec<f64>,
        min_proba: Option<f64>,
    ) -> bool { ... }
    fn is_transmitted(
        &self,
        frame_id: i32,
        scan_id: i32,
        mz: f64,
        min_proba: Option<f64>,
    ) -> bool { ... }
    fn any_transmitted(
        &self,
        frame_id: i32,
        scan_id: i32,
        mz: &Vec<f64>,
        min_proba: Option<f64>,
    ) -> bool { ... }
    fn transmit_tims_frame(
        &self,
        frame: &TimsFrame,
        min_probability: Option<f64>,
    ) -> TimsFrame { ... }
    fn transmit_tims_frame_annotated(
        &self,
        frame: &TimsFrameAnnotated,
        min_probability: Option<f64>,
    ) -> TimsFrameAnnotated { ... }
    fn isotopes_transmitted(
        &self,
        frame_id: i32,
        scan_id: i32,
        mz_mono: f64,
        isotopic_envelope: &Vec<f64>,
        min_probability: Option<f64>,
    ) -> (f64, Vec<(f64, f64)>) { ... }
}

Required Methods

fn apply_transmission( &self, frame_id: i32, scan_id: i32, mz: &Vec<f64>, ) -> Vec<f64>

Provided Methods

fn transmit_spectrum( &self, frame_id: i32, scan_id: i32, spectrum: MzSpectrum, min_probability: Option<f64>, ) -> MzSpectrum

Transmit a spectrum given a frame id and scan id

Arguments:

frame_id - frame id

scan_id - scan id

spectrum - MzSpectrum

min_probability - minimum probability for transmission

Returns:

MzSpectrum - transmitted spectrum

fn transmit_annotated_spectrum( &self, frame_id: i32, scan_id: i32, spectrum: MzSpectrumAnnotated, min_probability: Option<f64>, ) -> MzSpectrumAnnotated

Transmit an annotated spectrum given a frame id and scan id

Arguments:

frame_id - frame id

scan_id - scan id

spectrum - MzSpectrumAnnotated

min_probability - minimum probability for transmission

Returns:

MzSpectrumAnnotated - transmitted spectrum

fn transmit_ion( &self, frame_ids: Vec<i32>, scan_ids: Vec<i32>, spec: MzSpectrum, min_proba: Option<f64>, ) -> Vec<Vec<MzSpectrum>>

fn get_transmission_set( &self, frame_id: i32, scan_id: i32, mz: &Vec<f64>, min_proba: Option<f64>, ) -> HashSet<usize>

Get all ions in a frame that are transmitted

Arguments:

frame_id - frame id

scan_id - scan id

mz - mz values

min_proba - minimum probability for transmission

Returns:

HashSet<usize> - indices of transmitted mz values

fn all_transmitted( &self, frame_id: i32, scan_id: i32, mz: &Vec<f64>, min_proba: Option<f64>, ) -> bool

Check if all mz values in a given collection are transmitted

Arguments:

frame_id - frame id

scan_id - scan id

mz - mz values

min_proba - minimum probability for transmission

Returns:

bool - true if all mz values are transmitted

fn is_transmitted( &self, frame_id: i32, scan_id: i32, mz: f64, min_proba: Option<f64>, ) -> bool

Check if a single mz value is transmitted

Arguments:

frame_id - frame id

scan_id - scan id

mz - mz value

min_proba - minimum probability for transmission

Returns:

bool - true if mz value is transmitted

fn any_transmitted( &self, frame_id: i32, scan_id: i32, mz: &Vec<f64>, min_proba: Option<f64>, ) -> bool

Check if any mz value is transmitted, can be used to check if one peak of isotopic envelope is transmitted

Arguments:

frame_id - frame id

scan_id - scan id

mz - mz values

min_proba - minimum probability for transmission

Returns:

bool - true if any mz value is transmitted

fn transmit_tims_frame( &self, frame: &TimsFrame, min_probability: Option<f64>, ) -> TimsFrame

Transmit a frame given a diaPASEF transmission layout

fn transmit_tims_frame_annotated( &self, frame: &TimsFrameAnnotated, min_probability: Option<f64>, ) -> TimsFrameAnnotated

Transmit a frame given a diaPASEF transmission layout with annotations

Arguments:

frame - TimsFrameAnnotated

min_probability - minimum probability for transmission

Returns:

TimsFrameAnnotated - transmitted frame

fn isotopes_transmitted( &self, frame_id: i32, scan_id: i32, mz_mono: f64, isotopic_envelope: &Vec<f64>, min_probability: Option<f64>, ) -> (f64, Vec<(f64, f64)>)