mscore::algorithm

Module isotope

Functions

calculate_transmission_dependent_fragment_ion_isotope_distribution

convolve

convolve two distributions of masses and abundances

convolve_pow

convolve a distribution with itself n times

factorial

calculate the factorial of a number

generate_averagine_spectra

generate the averagine spectra for a given list of masses and charges using multiple threads

generate_averagine_spectrum

generate the averagine spectrum for a given mass and charge

generate_isotope_distribution

generate the isotope distribution for a given atomic composition

generate_isotope_pattern

generate the isotope pattern for a given mass and charge

generate_precursor_spectra

parallel version of generate_precursor_spectrum

generate_precursor_spectrum

generate the precursor spectrum for a given peptide sequence and charge using isotope convolutions

iso

calculate the isotope pattern for a given mass and charge based on the averagine model using the normal distribution for peak shapes

lam

calculate the lambda value for a given mass

normal_pdf

calculate the normal probability density function

weight