mscore::algorithm::isotope
Function iso

pub fn iso(
    x: &Vec<f64>,
    mass: f64,
    charge: f64,
    sigma: f64,
    amp: f64,
    k: usize,
    step_size: f64,
) -> Vec<f64>

calculate the isotope pattern for a given mass and charge based on the averagine model using the normal distribution for peak shapes

Arguments:

x - list of m/z values to probe

mass - mass of the peptide

charge - charge of the peptide

sigma - standard deviation of the normal distribution

amp - amplitude of the isotope pattern

k - number of isotopes to consider

step_size - step size for the m/z values to probe

Returns:

Vec<f64> - isotope pattern