mscore::algorithm::peptide

Function simulate_charge_states_for_sequences

pub fn simulate_charge_states_for_sequences(
    sequences: Vec<&str>,
    num_threads: usize,
    max_charge: Option<usize>,
    charged_probability: Option<f64>,
) -> Vec<Vec<f64>>

simulate the charge state distribution for a list of peptide sequences

Arguments

sequences - a vector of strings representing the peptide sequences

num_threads - an usize representing the number of threads to use

max_charge - an optional usize representing the maximum charge state to simulate

charged_probability - an optional f64 representing the probability of a site being charged

Returns

Vec<Vec<f64>> - a vector of vectors of f64 representing the probability of each charge state for each sequence

Example

use mscore::algorithm::peptide::simulate_charge_states_for_sequences;

let sequences = vec!["PEPTIDEH", "PEPTIDEH", "PEPTIDEH"];
let charge_state_probs = simulate_charge_states_for_sequences(sequences, 4, None, None);
assert_eq!(charge_state_probs, vec![vec![0.03999999999999999, 0.32, 0.64, 0.0, 0.0], vec![0.03999999999999999, 0.32, 0.64, 0.0, 0.0], vec![0.03999999999999999, 0.32, 0.64, 0.0, 0.0]]);